refactor(parser,review): lift dependency-graph.ts from review into parser
packages/review/src/dependency-graph.ts imported nothing review-specific --
only parser's own walkBounded, importMatchesTarget, normalizePath,
detectLanguage, findCSharpTypeReferenceDependents, and callerSymbolFor. Move
it to packages/parser/src/graph/dependency-graph.ts (alongside the generic
walkBounded BFS it already builds on) so its symbol/call-site-level caller
resolution is reachable from @liendev/cli too, not just @liendev/review --
a prerequisite for a richer get_dependents/lien api-delta blast-radius nudge
than today's file-level findDependents can give them.

Pure relocation: the moved file's only diff against its pre-move content is
the import block (package-level '@liendev/parser' imports rewritten to
parser-internal relative imports), verified with a byte diff against
origin/main. The moved test file gets a small local (non-exported)
createTestChunk factory in place of importing review's test-helpers.ts
across the package boundary, and one pre-existing unused parameter
(invisible under review's laxer tsconfig) is renamed with a leading
underscore to satisfy parser's noUnusedParameters -- no behavior change.

Review's 6 direct consumers, 5 test files, and 3 offline harness fixture
scripts that imported the local module now import from @liendev/parser
instead; review's own re-export of these symbols is deleted since nothing
outside review consumed it.
