+++ packages/review/test/blast-radius.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest';
 import type { CodeChunk } from '@liendev/parser';
-import { buildDependencyGraph } from '../src/dependency-graph.js';
+import { buildDependencyGraph } from '@liendev/parser';
 import { computeBlastRadius } from '../src/blast-radius.js';
 import { createTestChunk } from '../src/test-helpers.js';
